ai-service/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
anthropic==0.28.0
psycopg2-binary==2.9.9
redis==5.0.6
python-dotenv==1.0.1
pydantic==2.7.1
scikit-learn==1.5.0
numpy==1.26.4
pandas==2.2.2
langchain==0.2.5
langchain-anthropic==0.1.15
httpx==0.27.0
apscheduler==3.10.4
